requirements.txt
Flask
Flask-CORS
supabase
pyjwt
python-dotenv
SQLAlchemy
psycopg2-binary

